Backend_Daurin/requirements.txt
--index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://pypi.org/simple

torch==2.4.1+cpu

Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
gunicorn==22.0.0
Pillow==10.3.0
numpy==1.26.4
ultralytics==8.2.37
google-generativeai==0.7.2
psycopg2-binary==2.9.9
PyMySQL==1.1.0
